move async batch span processor to separate module
