+++ opentelemetry-sdk/Cargo.toml
@@ -56,6 +56,8 @@ internal-logs = ["tracing"]
 experimental_metrics_periodicreader_with_async_runtime = ["metrics"]
 spec_unstable_metrics_views = ["metrics"]
 experimental_logs_batch_log_processor_with_async_runtime = ["logs"]
+experimental_trace_batch_span_processor_with_async_runtime = ["trace"]
+
 
 [[bench]]
 name = "context"
